code

+++ Project/HotelReservation/src/reservations/PastReservationBean.java
@@ -35,7 +35,7 @@ public class PastReservationBean {
 		    
 		     while(rs.next()){ 
 		
-		    	 pastreservationss.add(new Reservation(rs.getString(1), rs.getDate(2), rs.getDate(3), rs.getInt(4), rs.getString(5),rs.getInt(6),rs.getDouble(7)));
+		    	 pastreservationss.add(new Reservation(rs.getString(1), rs.getDate(2), rs.getDate(3), rs.getInt(4), rs.getString(5),rs.getInt(6),rs.getInt(7)));
 	        }
 		} catch (Exception e) {
 			System.out.println(e.toString());
